makerequests adds requests to the api
