+++ front-end/src/App.js
@@ -8,32 +8,11 @@ import SignUp from "./SignUp"
 import Settings from './Settings'
 import Requests from './Requests'
 import MakeRequest from './MakeRequest'
-import mockData from './mockdata/requests.json'
 import EditInformation from './EditInformation';
-import {Box} from '@mui/material';
 const App = () => {
   // for adding items to a cart
   const [cartItems, setCartItems] = useState([]);
-  const showRequests = ({setCartItems, data}) => {
-    let list = data.map((val) => 
-      <div className="template" key={val.id}>
-          <Box className="description">
-            <div className = "itemControl"> Restaurant: <div className = "valueControl">{val.restaurant} </div></div>
-            <div className = "itemControl"> Party Size: <div className = "valueControl">{val.party_size}</div></div>
-            <div className = "itemControl"> Expiration Date: <div className = "valueControl">{val.expiration_date}</div></div>
-          </Box>
-      </div> 
-    )
-    setCartItems(list)
-  };
-  
-  const fetchRequests = async ({setCartItems}) => {
-      const response =  await fetch("https://635740569243cf412f954e2c.mockapi.io/api/rb/Request")
-      const data = await response.json()
-      showRequests({setCartItems,data})
-  }
 
-  
   return (
    <div className = "App">
     <Router> 
@@ -44,8 +23,8 @@ const App = () => {
               <Route path="/signup" element={<SignUp/>}/>
               <Route path="/settings" element={<Settings/>}/>
               <Route path = "/editinformation" element = {<EditInformation/>}/>
-              <Route path="/requests" element={<Requests cartItems = {cartItems} setCartItems = {setCartItems} fetchRequests = {fetchRequests}/>}/>
-              <Route path="/makerequest" element={<MakeRequest fetchRequests = {fetchRequests} />}/>
+              <Route path="/requests" element={<Requests cartItems = {cartItems} setCartItems = {setCartItems}/>}/>
+              <Route path="/makerequest" element={<MakeRequest />}/>
             </Routes>
           </Layout>
     </Router>
