+++ front-end/src/Requests.js
@@ -3,12 +3,25 @@ import React, { useEffect } from 'react'
 import mockData from './mockdata/requests.json'
 import {Box} from '@mui/material';
 
+const showRequests = ({setCartItems, data}) => {
+  let list = data.map((val) => 
+    <div className="template" key={val.id}>
+        <Box className="description">
+          <div className = "itemControl"> Restaurant: <div className = "valueControl">{val.restaurant} </div></div>
+          <div className = "itemControl"> Party Size: <div className = "valueControl">{val.party_size}</div></div>
+          <div className = "itemControl"> Expiration Date: <div className = "valueControl">{val.expiration_date}</div></div>
+        </Box>
+    </div> 
+  )
+  setCartItems(list)
+};
 
-
-
-
-
-const Requests = ({cartItems,setCartItems,fetchRequests}) => {
+const fetchRequests = async ({setCartItems}) => {
+    const response =  await fetch("https://635740569243cf412f954e2c.mockapi.io/api/rb/Request")
+    const data = await response.json()
+    showRequests({setCartItems,data})
+}
+const Requests = ({cartItems,setCartItems}) => {
     useEffect(()=>{fetchRequests({setCartItems})},[])
     return(
         <div>
